requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
sqlmodel>=0.0.14,<1.0.0
spnkr>=0.10.0,<1.0.0
aiohttp>=3.8.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
jinja2>=3.1.0,<4.0.0
aiofiles>=23.0.0,<25.0.0
